fixed major bug

+++ Relation.h
@@ -84,7 +84,8 @@ public:
         for(Tuple tuple : tuples){
             relation += "  ";
 
-            for(int col = 0; col < tuple.getValues().size() - 1 ; col++){
+            for (long unsigned int col = 0; col < tuple.getValues().size() - 1; col++)
+            {
                 relation += columnNames.getValues().at(col) + "=\'" + tuple.getValues().at(col) = "\', ";
             }
             relation += columnNames.getValues().at(tuple.getValues().size() - 1) + "=\'" + tuple.getValues().at(tuple.getValues().size() - 1) = "\'\n";
